fix s3 function crashes when given null arguments

+++ s3.js
@@ -14,31 +14,37 @@ const s3 = new S3({
 });
 
 exports.uploadFile = function(file) {
-  const fileStream = fs.createReadStream(file.path);
+  if (file) {
+    const fileStream = fs.createReadStream(file.path);
 
-  const uploadParams = {
-    Bucket: bucketName,
-    Body: fileStream,
-    Key: file.filename
-  };
+    const uploadParams = {
+      Bucket: bucketName,
+      Body: fileStream,
+      Key: file.filename
+    };
 
-  return s3.upload(uploadParams).promise();
+    return s3.upload(uploadParams).promise();
+  }
 };
 
 exports.getFileStream = function(fileKey) {
-  const downloadParams = {
-    Key: fileKey,
-    Bucket: bucketName
-  };
-
-  return s3.getObject(downloadParams).createReadStream();
+  if (fileKey) {
+    const downloadParams = {
+      Key: fileKey,
+      Bucket: bucketName
+    };
+  
+    return s3.getObject(downloadParams).createReadStream();
+  }
 };
 
 exports.deleteFile = function(fileKey) {
-  const deleteParams = {
-    Bucket: bucketName,
-    Key: fileKey
-  };
+  if (fileKey) {
+    const deleteParams = {
+      Bucket: bucketName,
+      Key: fileKey
+    };
 
-  return s3.deleteObject(deleteParams).promise();
+    return s3.deleteObject(deleteParams).promise();
+  }
 }
